test: clean up duplicated tests
