+++ test/integration/client-navigation/pages/head-with-json-ld-snippet.js
