--- test/integration/react-streaming-and-server-components/app/pages/runtime-rsc.server.js
@@ -1,9 +1,0 @@
-export default function () {
-  return process.version
-    ? `Runtime: Node.js ${process.version}`
-    : 'Runtime: Edge/Browser'
-}
-
-export const config = {
-  runtime: 'nodejs',
-}
