--- test/integration/react-streaming-and-server-components/app/pages/runtime.js
@@ -1,9 +1,0 @@
-export default function () {
-  return process.env.NEXT_RUNTIME === 'nodejs'
-    ? `Runtime: Node.js ${process.version}`
-    : 'Runtime: Edge/Browser'
-}
-
-export const config = {
-  runtime: 'nodejs',
-}
