+++ test/integration/react-streaming-and-server-components/test/streaming.js
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 import webdriver from 'next-webdriver'
-import { fetchViaHTTP, waitFor } from 'next-test-utils'
+import { fetchViaHTTP } from 'next-test-utils'
 import { getNodeBySelector } from './utils'
 
 async function resolveStreamResponse(response, onData) {
@@ -92,103 +92,6 @@ export default function (context, { env, runtime }) {
     expect(await browser.eval(`window.partial_hydration_counter_result`)).toBe(
       'count: 1'
     )
-  })
-
-  it('should flush the suffix at the very end', async () => {
-    await fetchViaHTTP(context.appPort, '/').then(async (response) => {
-      const result = await resolveStreamResponse(response)
-      expect(result).toMatch(/<\/body><\/html>/)
-    })
-  })
-
-  if (env === 'dev') {
-    it('should warn when stylesheets or scripts are in head', async () => {
-      let browser
-      try {
-        browser = await webdriver(context.appPort, '/head')
-
-        await browser.waitForElementByCss('h1')
-        await waitFor(1000)
-        const browserLogs = await browser.log('browser')
-        let foundStyles = false
-        let foundScripts = false
-        const logs = []
-        browserLogs.forEach(({ message }) => {
-          if (message.includes('Do not add stylesheets using next/head')) {
-            foundStyles = true
-            logs.push(message)
-          }
-          if (message.includes('Do not add <script> tags using next/head')) {
-            foundScripts = true
-            logs.push(message)
-          }
-        })
-
-        expect(foundStyles).toEqual(true)
-        expect(foundScripts).toEqual(true)
-
-        // Warnings are unique
-        expect(logs.length).toEqual(new Set(logs).size)
-      } finally {
-        if (browser) {
-          await browser.close()
-        }
-      }
-    })
-
-    it('should warn when scripts are in head', async () => {
-      let browser
-      try {
-        browser = await webdriver(context.appPort, '/head')
-        await browser.waitForElementByCss('h1')
-        await waitFor(1000)
-        const browserLogs = await browser.log('browser')
-        let found = false
-        browserLogs.forEach((log) => {
-          if (log.message.includes('Use next/script instead')) {
-            found = true
-          }
-        })
-        expect(found).toEqual(true)
-      } finally {
-        if (browser) {
-          await browser.close()
-        }
-      }
-    })
-
-    it('should not warn when application/ld+json scripts are in head', async () => {
-      let browser
-      try {
-        browser = await webdriver(context.appPort, '/head-with-json-ld-snippet')
-        await browser.waitForElementByCss('h1')
-        await waitFor(1000)
-        const browserLogs = await browser.log('browser')
-        let found = false
-        browserLogs.forEach((log) => {
-          if (log.message.includes('Use next/script instead')) {
-            found = true
-          }
-        })
-        expect(found).toEqual(false)
-      } finally {
-        if (browser) {
-          await browser.close()
-        }
-      }
-    })
-  }
-
-  it('should stream to users', async () => {
-    const res = await fetchViaHTTP(context.appPort, '/streaming')
-    let flushCount = 0
-    await resolveStreamResponse(res, () => {
-      flushCount++
-    })
-    expect(flushCount).toBeGreaterThan(1)
-    if (runtime === 'nodejs') {
-      expect(res.headers.get('etag')).toBeNull()
-    }
   })
 
   it('should not stream to crawlers or google pagerender bot', async () => {
